PR for Forms into Master

Related work items: #37240, #37276, #37278, #37279, #37280, #37304, #38158, #38177, #38303, #38537

+++ ApolloQA/Pages/Quote/Quote_BusinessInformation_Page.cs
@@ -6,8 +6,8 @@ namespace ApolloQA.Pages
     {
         public static string URL_PATH = "/quote/{quoteID}/section/{sectionID}";
 
-        public static string GetURL(int quoteID, long sectionID) => Quote_Page.GetURL(quoteID, sectionID);
-        public static void Navigate(int quoteID, long sectionID) => Quote_Page.Navigate(quoteID, sectionID);
+        public static string GetURL(long quoteID, long sectionID) => Quote_Page.GetURL(quoteID, sectionID);
+        public static void Navigate(long quoteID, long sectionID) => Quote_Page.Navigate(quoteID, sectionID);
 
         public static Element Score => new Element("//mat-card-title[text()='National Credit File']/following-sibling::mat-card-content[contains(text(), 'Score:')]");
